backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
asyncpg==0.30.0
sqlalchemy[asyncio]==2.0.36
alembic==1.14.1
pydantic-settings==2.7.1
elasticsearch[async]==8.17.0
redis[hiredis]==5.2.1
httpx==0.28.1
